Add data to the package
Add data to the package

I added a data set, 'pop', that will be useful for examples and tests. The script to make the data is in data-raw. The data set is in data/ and the dataset is documented in R/data.R.

I've also added a second data set, samp_srcs which is a simple random cluster sample of the pop data