Add some tests and objects for tests
This is a test of tests. I've added three tests that should fail if the output from coef, print, or summary changes unexpectedly. the reference objects are created in data-raw\generate_objects_for_tests.R and use the sample dataset samp_srcs.

I also fixed a bug with print. It was throwing an error when there was one fixed effect in the model because of an inadvertent dropping of the matrix class. Fixed by adding drop=FALSE to the offending matrix sub setting.