Debug residuals.coxme
This commit corrects two issues with residuals.coxme. Tied times had the wrong residuals, and residuals were produced when event times were 0, when they should have been missing.

The first issues was a bug in agscore3.cpp, introduced when I ported it from C, where the index i was not updated correctly part way through the script.

The second issue was to do with using agscore3 with right censored time. The function was ignoring when event times were 0, when this should be an error.

I corrected the bugs in agscore3.cpp and added the function coxscore2 to handle right-censored times.