fix broken roxygen tags and tests
I'm still figuring out where namespaces need to be used, and the right roxygen tags to use. I had removed @export from the methods, exporting only the generic and that was wrong, so now the export tags are back and the methods are linked to the generic documentation.

For tests, I needed to remove a namespace, survival::Surv. I hope that's what I needed to do anyway. Will see if the tests work now.